% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env}
\alias{env}
\title{Get environmental data}
\usage{
env()
}
\description{
Download or describe past downloads of environmental data
}
\details{
#' @examples
\dontrun{
       x <- y
       # put example here
    }
}
