% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norway.R
\name{norway}
\alias{norway}
\title{Create an outline of Norway}
\usage{
norway(lonlat = TRUE)
}
\arguments{
\item{lonlat}{Should the funtion return a raster in the lonlat coordinate system (defult) or not, in which case it return it as UTM32.}
}
\value{
Returns a Large SpatialPolygonsDataFrame
}
\description{
\code{norway} returs the outline of Norway
}
\details{
Here comes the details...
}
\examples{
\dontrun{
norway() # for a lonlat polygon
norway(lonlat = F) # for a UTM32 polygon
}
}
