% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard error of the mean}
\usage{
se(x)
}
\description{
Calculates SE
}
\examples{
\dontrun{
# craete a vector with sd = 1 and n = 150. SE should be 1/sqrt(100) = 0.08
a <- rnorm(150,5,1)
se() 

# it automatically removed NA's
a2 <- a
a2[3] <- NA
se(a2)
}
}
