% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spList.R
\name{sl}
\alias{sl}
\title{Get species lists}
\usage{
sl(df = FALSE)
}
\arguments{
\item{df}{Logical. Whether to return a dataframe which also includes the species group in a second column. Default is df = FALSE which returns acharacter string of names (type: vector).}
}
\description{
This scripts produces a list of the species that we will quiery gbif for occurence data.
\code{sl()} produces a species list and a dataframe with the same list pluss some more
}
\examples{
\dontrun{
       myList <- sl()
    }
}
